Working Button Concept
I've managed to create a functioning prototype for this button I've been
planning to make for the past few days. The concept is to have a circle
that, when a user hovers over, triggers a series of animations to pop up.
All objects were centered using position absolute, and animations were
done using keyframe animation.

My next step is to figure out how I can create proportional SVG icons
that I can incorporate into the design. The idea for this particular
button is to have a small town at the top pop up from left to right,
then have some natural objects--such as trees and mountains-- pop
up on the bottom. Then, as the city animates from left to right a
train will animate and circle the entire button. I think I might be
able to add a smoke stack effect to the train if I can animate small
particles using opacity... If I try to add the smoke as part of the
train's animation then it will look off because the smoke will follow
the train instead of being left behind.

More to come tomorrow. I'm very excited for this one.