Added Train Animation
Train animation finally works. It uses two copies of the same
SVG train icon. The first one is used to animate the pop-up
effect. Once the pop-up effect is completed, the first image
fades away. The second image, which is in the same location as
the first image, then begins the rotation animation.

The reason I needed 2 SVGs is because I wanted the train to keep
looping around the circle while the mouse is hovered over the
circle. If I had to combine the pop-up animation with the rotation
animation, then the train would do a small pop-up each time I
looped the animation.

Next section will be adding in the other items that appear as
the train makes its rotation.