Second Car Animation Added
I ran into an issue with the rotation of adding a second car. The
rotation always wanted to go counter clockwise. However, I found that
the browser will allow you to use numbers above 360 and below 0. This
means that you can control the direction of a rotation, so long as you
think in terms of 360 degrees. For my example my second train car starts
at 335 degrees and needs to make a 360 rotation. So my start point begins
at 335 and ends at 695, which is 335+360. This is super helpful because
no matter what I did I couldn't seem to control the rotation of the cars.