Added pop-in building animations
I honestly thought this would be simpler than it was. I added the
building fade in animation. Again I had to break this into two
parts: Part I was animating the image in, and Part II was keeping
the image static once the animation was finished. I achieved this
by creating two different animations: one for the fade in and
the other to maintain the opacity while the user hovers. Now
I should be able to replicate this effect for the other SVG images
and get close to finishing this effect.