Working Project Version 1
The hover effect is finally complete. The SVGs I made now
animate in on a hover and stay while the train passes by.
I'm not 100% satisfied with the final design of the effect.
I feel like it's not balanced, so I might add in a few other
things, maybe clouds/trees, to balance out some of the white
space. But, overall I'm happy this is finally working.