Working Project Version 1.1
I was able to add cloud animations that bring the world
I was attmpting to make together. I also changed the image
to a button with text that, when clicked, will navigatge to
the Wikipedia page on France. For now I think I'm going to
stop working on this button, but I feel like I've been able
to achieve what I wanted to do: make a small/miniature world
that was also a button. I wanted to make a button that told
a story of where you were going to go before you clicked on
it. I think the things I learned from making this small project
will help me as I now have a few more ideas for my own personal
website.